resume-agent/requirements.txt
crewai
langchain
langchainhub
langchain-community
groq
pymupdf
python-dotenv
tqdm
langchain-groq
fastapi
python-multipart
